Compat: add `open` and `os` shims
