+++ pathlib2/compat/constants.py
@@ -1,0 +1,75 @@
+
+# arranged according to the documentation here https://docs.python.org/2/library/os.html
+OS_FNS_TAKING_PATHS = {
+    # Files and Directories
+    'access',
+    'chdir',
+    'chroot',
+    'chmod',
+    'chflags',
+    'chroot',
+    'chown',
+    'lchflags',
+    'lchmod',
+    'lchown',
+    'link',
+    'listdir',
+    'lstat',
+    'mkfifo',
+    'mknod',
+    'mkdir',
+    'makedirs',
+    'pathconf',
+    'remove',
+    'removedirs',
+    'rename',
+    'renames',
+    'rmdir',
+    'stat',
+    'statvfs',
+    'symlink',
+    'tempnam',
+    'utime',
+    'walk',
+    # Process Management
+    'execl',
+    'execle',
+    'execv',
+    'execve',
+    'spawnl',
+    'spawnle',
+    'spawnv',
+    'spawnve',
+    'startfile'
+}
+
+# arranged according to the documentation here https://docs.python.org/2/library/os.path.html
+OS_PATH_FNS_TAKING_PATHS = {
+    'abspath',
+    'basename',
+    'dirname',
+    'exists',
+    'lexists',
+    'expanduser',
+    'expandvars',
+    'getatime',
+    'getmtime',
+    'getctime',
+    'getsize',
+    'isabs',
+    'isfile',
+    'isdir',
+    'islink',
+    'ismount',
+    'join',
+    'normcase',
+    'normpath',
+    'realpath',
+    'relpath',
+    'samefile',
+    'split',
+    'splitdrive',
+    'splitext',
+    'splitunc',
+    'walk',
+}
