+++ requirements.txt
@@ -1,3 +1,4 @@
 six
 scandir; python_version < '3.5'
 mock; python_version < '3.3'
+pytest
