+++ tests/test_pathlib2.py
@@ -6,7 +6,6 @@
 import io
 import os
 import errno
-from .context import pathlib2 as pathlib
 import pickle
 import six
 import socket
@@ -14,89 +13,11 @@ import stat
 import sys
 import tempfile
 
-if sys.version_info >= (3, 3):
-    import collections.abc as collections_abc
-else:
-    import collections as collections_abc
-
-if sys.version_info < (2, 7):
-    try:
-        import unittest2 as unittest
-    except ImportError:
-        raise ImportError("unittest2 is required for tests on pre-2.7")
-else:
-    import unittest
-
-if sys.version_info < (3, 3):
-    try:
-        import mock
-    except ImportError:
-        raise ImportError("mock is required for tests on pre-3.3")
-else:
-    from unittest import mock
-
-# assertRaisesRegex is missing prior to Python 3.2
-if sys.version_info < (3, 2):
-    unittest.TestCase.assertRaisesRegex = unittest.TestCase.assertRaisesRegexp
-
-try:
-    from test import support
-except ImportError:
-    from test import test_support as support
-
-if sys.version_info >= (3, 6):
-    android_not_root = support.android_not_root
-else:
-    android_not_root = False
-
-TESTFN = support.TESTFN
-
-# work around broken support.rmtree on Python 3.3 on Windows
-if (os.name == 'nt'
-        and sys.version_info >= (3, 0) and sys.version_info < (3, 4)):
-    import shutil
-    support.rmtree = shutil.rmtree
-
-try:
-    import grp
-    import pwd
-except ImportError:
-    grp = pwd = None
-
-# support.can_symlink is missing prior to Python 3
-if six.PY2:
-
-    def support_can_symlink():
-        return pathlib.supports_symlinks
-
-    support_skip_unless_symlink = unittest.skipIf(
-        not pathlib.supports_symlinks,
-        "symlinks not supported on this platform")
-else:
-    support_can_symlink = support.can_symlink
-    support_skip_unless_symlink = support.skip_unless_symlink
-
-
-# Backported from 3.4
-def fs_is_case_insensitive(directory):
-    """Detects if the file system for the specified directory is
-    case-insensitive.
-    """
-    base_fp, base_path = tempfile.mkstemp(dir=directory)
-    case_path = base_path.upper()
-    if case_path == base_path:
-        case_path = base_path.lower()
-    try:
-        return os.path.samefile(base_path, case_path)
-    except OSError as e:
-        if e.errno != errno.ENOENT:
-            raise
-        return False
-    finally:
-        os.unlink(base_path)
-
-
-support.fs_is_case_insensitive = fs_is_case_insensitive
+from .context import (
+    pathlib2 as pathlib,
+    collections_abc, unittest, mock, support, android_not_root, TESTFN, grp, pwd, support_can_symlink,
+    support_skip_unless_symlink
+)
 
 
 class _BaseFlavourTest(object):
@@ -1512,7 +1433,8 @@ class _BasePathTest(object):
         # check that trying to write str does not truncate the file
         with self.assertRaises(TypeError) as cm:
             (p / 'fileA').write_bytes(six.u('somestr'))
-        self.assertTrue(str(cm.exception).startswith('data must be'))
+        msg = str(cm.exception)
+        self.assertTrue(msg.startswith('data must be') or msg.startswith('memoryview: a bytes-like object'))
         self.assertEqual((p / 'fileA').read_bytes(), b'abcdefg')
 
     def test_read_write_text(self):
@@ -1987,7 +1909,7 @@ class _BasePathTest(object):
             def _else_func():
                 self.assertNotIn(str(p12), concurrently_created)
 
-            pathlib._try_except_fileexistserror(
+            pathlib.utils._try_except_fileexistserror(
                 _try_func, _exc_func, _else_func)
             self.assertTrue(p.exists())
